Dynamically calculate summary line position
Summary line position is no longer fixed in rspec. When specs run
successfully it is at -4, but if something fails, the failing tests are
listed after the summary, making impossible to establish a predefined
summary position. Due to this, I'm using the summary regex to find the
summary line, no longer needing a predefined summary position.